provide Article as lineItems via OrderRequest and PatchRequest
The feature existed as commented out parked code.
That's why it has been reactivated.
The reason for this is that we generally no longer need to transfer taxes.
The calculation of taxes in Netto shops caused errors on PayPal's side.
We only need to transfer taxes in the OrderRequest from PUI.
Since PUI is only allowed for B2C, we can transfer the taxes here.
Since PUI does not need a patch request, the taxes are not included in
the PatchRequest, but only in the OrderRequest if necessary.